Refactoring of WebDAV code
