+++ k9mail-library/src/main/java/com/fsck/k9/mail/store/webdav/ConnectionInfo.java
@@ -1,0 +1,11 @@
+package com.fsck.k9.mail.store.webdav;
+
+/**
+ * Simple data container for passing connection information.
+ */
+class ConnectionInfo {
+    public int statusCode;
+    public short requiredAuthType;
+    public String guessedAuthUrl;
+    public String redirectUrl;
+}
